dom/performance/tests/test_worker_user_timing.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
  <head>
    <title>Test for worker performance timing API</title>
    <meta http-equiv="content-type" content="text/html; charset=UTF-8">
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
  </head>
  <body>
    <script class="testbody" type="text/javascript">

var worker = new Worker('worker_performance_user_timing.js');
worker.onmessage = function(event) {
  if (event.data.type == 'finish') {
    SpecialPowers.setBoolPref("privacy.reduceTimerPrecision", reduceTimePrecisionPrevPrefValue);
    SimpleTest.finish();
  } else if (event.data.type == 'status') {
    ok(event.data.status, event.data.msg);
  }
}

var reduceTimePrecisionPrevPrefValue = SpecialPowers.getBoolPref("privacy.reduceTimerPrecision");
SpecialPowers.setBoolPref("privacy.reduceTimerPrecision", false);
SimpleTest.waitForExplicitFinish();
    </script>
  </body>
</html>
